Update plugin temalpte to reflect migration guide
I've changed the example plugins file structure to reflect https://github.com/elastic/kibana/pull/31147/files
but didn't change any of the parameter names or interfaces where these differ from the guide.
